backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
reportlab
python-dotenv